+++ src/service/judging.test.ts
@@ -1,9 +1,9 @@
 import type { EmbedMessage } from '../model/embed-message';
-import { JudgementCommand } from './judgement';
+import { JudgingCommand } from './judging';
 import { createMockMessage } from './command-message';
 
 test('use case of jd', async () => {
-  const responder = new JudgementCommand({
+  const responder = new JudgingCommand({
     sleep: () => Promise.resolve(),
     uniform: () => 2
   });
@@ -17,7 +17,7 @@ test('use case of jd', async () => {
       },
       (embed) => {
         expect(embed).toStrictEqual({
-          title: '***†HARACHO ONLINE JUDGEMENT SYSTEM†***',
+          title: '***†HARACHO ONLINE JUDGING SYSTEM†***',
           description: '0 / 5 WJ'
         });
         return Promise.resolve({ edit: fn });
@@ -28,18 +28,18 @@ test('use case of jd', async () => {
   expect(fn).toBeCalledTimes(5);
   for (let i = 0; i < 4; ++i) {
     expect(fn.mock.calls[i][0]).toStrictEqual({
-      title: '***†HARACHO ONLINE JUDGEMENT SYSTEM†***',
+      title: '***†HARACHO ONLINE JUDGING SYSTEM†***',
       description: `${i + 1} / 5 WJ`
     });
   }
   expect(fn.mock.calls[4][0]).toStrictEqual({
-    title: '***†HARACHO ONLINE JUDGEMENT SYSTEM†***',
+    title: '***†HARACHO ONLINE JUDGING SYSTEM†***',
     description: `5 / 5 AC`
   });
 });
 
 test('use case of judge', async () => {
-  const responder = new JudgementCommand({
+  const responder = new JudgingCommand({
     sleep: () => Promise.resolve(),
     uniform: () => 0
   });
@@ -53,7 +53,7 @@ test('use case of judge', async () => {
       },
       (embed) => {
         expect(embed).toStrictEqual({
-          title: '***†HARACHO ONLINE JUDGEMENT SYSTEM†***',
+          title: '***†HARACHO ONLINE JUDGING SYSTEM†***',
           description: '0 / 1 WJ'
         });
         return Promise.resolve({ edit: fn });
@@ -63,13 +63,13 @@ test('use case of judge', async () => {
 
   expect(fn).toBeCalledTimes(1);
   expect(fn.mock.calls[0][0]).toStrictEqual({
-    title: '***†HARACHO ONLINE JUDGEMENT SYSTEM†***',
+    title: '***†HARACHO ONLINE JUDGING SYSTEM†***',
     description: `1 / 1 WA`
   });
 });
 
 test('max number of cases', async () => {
-  const responder = new JudgementCommand({
+  const responder = new JudgingCommand({
     sleep: () => Promise.resolve(),
     uniform: () => 1
   });
@@ -82,7 +82,7 @@ test('max number of cases', async () => {
       },
       (embed) => {
         expect(embed).toStrictEqual({
-          title: '***†HARACHO ONLINE JUDGEMENT SYSTEM†***',
+          title: '***†HARACHO ONLINE JUDGING SYSTEM†***',
           description: '0 / 1 WJ'
         });
         return Promise.resolve({ edit: () => Promise.resolve() });
@@ -97,7 +97,7 @@ test('max number of cases', async () => {
       },
       (embed) => {
         expect(embed).toStrictEqual({
-          title: '***†HARACHO ONLINE JUDGEMENT SYSTEM†***',
+          title: '***†HARACHO ONLINE JUDGING SYSTEM†***',
           description: '0 / 64 WJ'
         });
         return Promise.resolve({ edit: () => Promise.resolve() });
