+++ src/service/judging.ts
@@ -6,7 +6,7 @@ import type {
 import type { MessageEvent } from '../runner';
 
 /**
- * `JudgementCommand` のための乱数生成器。
+ * `JudgingCommand` のための乱数生成器。
  *
  * @export
  * @interface RandomGenerator
@@ -31,18 +31,18 @@ export interface RandomGenerator {
   uniform(from: number, to: number): number;
 }
 
-const JUDGEMENT_TITLE = '***†HARACHO ONLINE JUDGEMENT SYSTEM†***';
+const JUDGING_TITLE = '***†HARACHO ONLINE JUDGING SYSTEM†***';
 
 /**
  * `judge` コマンドで競技プログラミングの判定をシミュレートする。
  *
  * @export
- * @class JudgementCommand
+ * @class JudgingCommand
  * @implements {MessageEventResponder<CommandMessage>}
  */
-export class JudgementCommand implements CommandResponder {
+export class JudgingCommand implements CommandResponder {
   help: Readonly<HelpInfo> = {
-    title: JUDGEMENT_TITLE,
+    title: JUDGING_TITLE,
     description: 'プログラムが適格かどうか判定してあげるよ',
     commandName: ['jd', 'judge'],
     argsFormat: [
@@ -89,19 +89,19 @@ export class JudgementCommand implements CommandResponder {
 
   private async accept(message: CommandMessage, count: number) {
     const sent = await message.reply({
-      title: JUDGEMENT_TITLE,
+      title: JUDGING_TITLE,
       description: `0 / ${count} WJ`
     });
 
     for (let i = 1; i <= count - 1; ++i) {
       await sent.edit({
-        title: JUDGEMENT_TITLE,
+        title: JUDGING_TITLE,
         description: `${i} / ${count} WJ`
       });
       await this.rng.sleep();
     }
     await sent.edit({
-      title: JUDGEMENT_TITLE,
+      title: JUDGING_TITLE,
       description: `${count} / ${count} AC`
     });
   }
@@ -113,7 +113,7 @@ export class JudgementCommand implements CommandResponder {
     result: string
   ) {
     const sent = await message.reply({
-      title: JUDGEMENT_TITLE,
+      title: JUDGING_TITLE,
       description: `0 / ${count} WJ`
     });
 
@@ -122,13 +122,13 @@ export class JudgementCommand implements CommandResponder {
 
     for (let i = 1; i <= count - 1; ++i) {
       await sent.edit({
-        title: JUDGEMENT_TITLE,
+        title: JUDGING_TITLE,
         description: `${i} / ${count} ${errorAt <= i ? result : 'WJ'}`
       });
       await this.rng.sleep();
     }
     await sent.edit({
-      title: JUDGEMENT_TITLE,
+      title: JUDGING_TITLE,
       description: `${count} / ${count} ${result}`
     });
   }
